[refactor] 동기화 Usecase 네이밍 변경
