+++ order-query-api/order-query-application/src/main/java/org/flab/deliveryplatform/order/query/application/port/AddMyOrderUseCase.java
@@ -1,0 +1,9 @@
+package org.flab.deliveryplatform.order.query.application.port;
+
+import org.flab.deliveryplatform.order.query.application.port.dto.AddMyOrderCommand;
+
+public interface AddMyOrderUseCase {
+
+    void addMyOrder(AddMyOrderCommand command);
+
+}
