+++ order-query-api/order-query-application/src/main/java/org/flab/deliveryplatform/order/query/application/port/UpdateMyOrderUseCase.java
@@ -1,0 +1,8 @@
+package org.flab.deliveryplatform.order.query.application.port;
+
+public interface UpdateMyOrderUseCase {
+    
+    void updateMyOrderStatus(Long orderId, String status);
+
+    void updateMyOrderDeliveryStatus(Long orderId, String deliveryStatus);
+}
