+++ order-query-api/order-query-application/src/main/java/org/flab/deliveryplatform/order/query/application/service/AddMyOrderService.java
@@ -2,11 +2,10 @@ package org.flab.deliveryplatform.order.query.application.service;
 
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
+import org.flab.deliveryplatform.order.query.application.port.AddMyOrderUseCase;
 import org.flab.deliveryplatform.order.query.application.port.GetShopRepository;
 import org.flab.deliveryplatform.order.query.application.port.MyOrderRepository;
-import org.flab.deliveryplatform.order.query.application.port.SyncMyOrderUseCase;
-import org.flab.deliveryplatform.order.query.application.port.dto.SyncMyOrderCommand;
-import org.flab.deliveryplatform.order.query.application.port.exception.MyOrderNotFoundException;
+import org.flab.deliveryplatform.order.query.application.port.dto.AddMyOrderCommand;
 import org.flab.deliveryplatform.order.query.domain.MyOrder;
 import org.flab.deliveryplatform.order.query.domain.MyOrder.MyOrderLineItem;
 import org.flab.deliveryplatform.order.query.domain.Shop;
@@ -14,9 +13,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @RequiredArgsConstructor
-@Transactional(readOnly = true)
 @Service
-public class SyncMyOrderService implements SyncMyOrderUseCase {
+public class AddMyOrderService implements AddMyOrderUseCase {
 
     private final MyOrderRepository myOrderRepository;
 
@@ -24,7 +22,7 @@ public class SyncMyOrderService implements SyncMyOrderUseCase {
 
     @Transactional
     @Override
-    public void syncMyOrder(SyncMyOrderCommand command) {
+    public void addMyOrder(AddMyOrderCommand command) {
         Shop shop = getShopRepository.findById(command.getShopId())
             .orElseThrow(() -> new IllegalArgumentException("가게가 존재하지 않습니다."));
 
@@ -42,24 +40,5 @@ public class SyncMyOrderService implements SyncMyOrderUseCase {
             .build();
 
         myOrderRepository.save(myOrder);
-    }
-
-    @Transactional
-    @Override
-    public void syncMyOrderStatus(Long orderId, String status) {
-        MyOrder myOrder = findMyOrder(orderId);
-        myOrder.changeStatus(status);
-    }
-
-    @Transactional
-    @Override
-    public void syncMyOrderDeliveryStatus(Long orderId, String deliveryStatus) {
-        MyOrder myOrder = findMyOrder(orderId);
-        myOrder.changeDeliveryStatus(deliveryStatus);
-    }
-
-    private MyOrder findMyOrder(Long orderId) {
-        return myOrderRepository.findByOrderId(orderId)
-            .orElseThrow(() -> new MyOrderNotFoundException(orderId));
     }
 }
