+++ order-query-api/order-query-interfaces/src/main/java/org/flab/deliveryplatform/order/query/interfaces/eventlistener/OrderDeliveredApplicationEventListener.java
@@ -1,7 +1,7 @@
 package org.flab.deliveryplatform.order.query.interfaces.eventlistener;
 
 import lombok.RequiredArgsConstructor;
-import org.flab.deliveryplatform.order.query.application.port.SyncMyOrderUseCase;
+import org.flab.deliveryplatform.order.query.application.port.UpdateMyOrderUseCase;
 import org.flab.deliveryplatform.order.query.interfaces.eventlistener.event.OrderDeliveredApplicationEvent;
 import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
@@ -10,10 +10,10 @@ import org.springframework.stereotype.Component;
 @Component
 public class OrderDeliveredApplicationEventListener {
 
-    private final SyncMyOrderUseCase syncMyOrderUseCase;
+    private final UpdateMyOrderUseCase updateMyOrderUseCase;
 
     @EventListener
     public void handle(OrderDeliveredApplicationEvent event) {
-        syncMyOrderUseCase.syncMyOrderStatus(event.getOrderId(), event.getStatus());
+        updateMyOrderUseCase.updateMyOrderStatus(event.getOrderId(), event.getStatus());
     }
 }
